change: ログイン画面を分割し、アプリ起動後画面をホーム画面へ変更
