+++ src/lib/main.dart
@@ -1,11 +1,11 @@
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_authenticator/amplify_authenticator.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter/material.dart';
+import 'package:yamabico/feature/auth/presentation/login_screen.dart';
+import 'package:yamabico/home_screen.dart';
 
 import 'feature/auth/infra/amplifyconfiguration.dart';
-import 'feature/auth/presentation/amplify_cognito_resolver.dart';
 
 void main() async {
   await dotenv.load(fileName: '.env');
@@ -39,76 +39,12 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return Authenticator(
-      stringResolver: stringResolver,
-      authenticatorBuilder: (BuildContext context, AuthenticatorState state) {
-        switch (state.currentStep) {
-          case AuthenticatorStep.signIn:
-            return CustomScaffold(
-              state: state,
-              body: SignInForm(),
-            );
-          case AuthenticatorStep.resetPassword:
-            return CustomScaffold(
-              state: state,
-              body: ResetPasswordForm(),
-            );
-          case AuthenticatorStep.confirmResetPassword:
-            return CustomScaffold(
-              state: state,
-              body: const ConfirmResetPasswordForm(),
-            );
-          default:
-            return null;
-        }
+    return MaterialApp(
+      initialRoute: '/',
+      routes: {
+        '/': (context) => const HomeScreen(),
+        '/login': (context) => const LoginScreen(),
       },
-      child: MaterialApp(
-        builder: Authenticator.builder(),
-        home: Scaffold(
-          appBar: AppBar(title: const Text('ホーム')),
-          body: const Center(
-            child: Text('ログイン成功！', style: TextStyle(fontSize: 32.0)),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class CustomScaffold extends StatelessWidget {
-  const CustomScaffold({
-    super.key,
-    required this.state,
-    required this.body,
-    this.footer,
-  });
-
-  final AuthenticatorState state;
-  final Widget body;
-  final Widget? footer;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Padding(
-        padding: const EdgeInsets.all(16),
-        child: SingleChildScrollView(
-          child: Column(
-            children: [
-              // App logo
-              const Padding(
-                padding: EdgeInsets.only(top: 32),
-                child: Center(child: FlutterLogo(size: 100)),
-              ),
-              Container(
-                constraints: const BoxConstraints(maxWidth: 600),
-                child: body,
-              ),
-            ],
-          ),
-        ),
-      ),
-      persistentFooterButtons: footer != null ? [footer!] : null,
     );
   }
 }
